Introduce IDoor
